debug xover
Problem was that t12a was able to identify a graph where the search for
symmetric subgraphs could find a combination of symmetric sites where one symmetric subgraph had 4 vertexes, 3 of which CAPs, and the other had 4 vertexes, all proper vertexes, nd the roots of these symmetric sites were indeed symmetris (while the other 3 vertex were not). Attempt to do crossover would remove the caps in one site, thus leading to the identification of a single-vertex symmetric subgraph on either side. This resulted in crossover operation properly done on the correct one-vertex site (after removal of CAPs), but the attempt to do it also on the symmetric side would find used APs without mapping due to the 3 child fragment not being CAPs and, therefore, not being removed from the graph, while their parent vertex was expected to be swapped by a single-AP incoming graph. This is not possible, indeed.
Problem fixed by checking that the symmetric sites do retain consistency in the first shell of vertex surrounding the swappable subgraph.